big data/requirements.txt.txt
tensorflow-cpu
nltk
flask
kafka-python
pyspark==3.2.1